# openapi.yaml
components:
  schemas:
...
    http.MimeMismatchDetail:
      properties:
        detectedMime:
          type: string
        knownMime:
          type: string
      required:
        - knownMime
        - detectedMime
      type: object
    http.RejectedContentResponse:
      properties:
        code:
          type: string
        detail:
          $ref: '#/components/schemas/http.MimeMismatchDetail'
        error:
          type: string
      required:
        - code
        - error
      type: object
    http.ReplaceContentResponse:
      properties:
        externalID:
          type: string
        imageHeight:
          type: integer
        imageWidth:
          type: integer
        mimeType:
          type: string
        size:
          type: integer
      required:
        - externalID
        - mimeType
        - size
      type: object
...
    http.errorResponse:
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
      type: object
info:
  contact:
    email: info@alkem.io
    name: Alkemio Foundation
    url: https://alkem.io
  description: |
    Universal file I/O gateway and document lifecycle manager for the Alkemio platform.
    Provides authenticated file serving (public), document CRUD, image processing,
    and MIME validation (internal).
  license:
    name: EUPL-1.2
    url: https://opensource.org/licenses/EUPL-1.2
  title: Alkemio File Service
  version: 0.2.0
openapi: 3.1.1
paths:
  /internal/blob/{hash}/content:
    get:
      description: |-
        GetBlobContent handles GET /internal/blob/{hash}/content — a content-addressed
        read. It returns the immutable blob stored under a SHA3-256 content hash,
        independent of any document row. The store is already content-addressed
        (Storage.Read keys on the hash); this exposes that read over the API.

        This is the read path the backup/replication worker uses (workspace#008,
        FR-008): the outbox records an object's content hash, and the blob under a
        hash never changes, so a fetch-by-hash is version-exact — it returns the
        EXACT bytes that were enqueued, never whichever version the (mutable)
        document currently points at, which is what GET /file/{id}/content returns.
        A hash with no blob (e.g. a version superseded and refcount-deleted) is a
        clean 404, distinguishable by the caller from a transport failure.
      operationId: DocumentHandler.GetBlobContent
      parameters:
        - in: path
          name: hash
          required: true
          schema:
            type: string
      responses:
        "200":
          content:
            application/octet-stream:
              schema:
                format: binary
                type: string
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/http.errorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/http.errorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/http.errorResponse'
          description: Internal Server Error
      summary: |-
        GetBlobContent handles GET /internal/blob/{hash}/content — a content-addressed
        read.
      tags:
        - /internal
...
  /internal/file/{id}/content:
    get:
      operationId: DocumentHandler.GetContent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: uuid
            type: string
      responses:
        "200":
          content:
            application/octet-stream:
              schema:
                format: binary
                type: string
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/http.errorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/http.errorResponse'
          description: Internal Server Error
      summary: GetContent handles GET /internal/file/{id}/content
      tags:
        - /internal
    put:
      operationId: DocumentHandler.ReplaceContent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: uuid
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/http.ReplaceContentResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/http.errorResponse'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/http.errorResponse'
          description: Not Found
        "408":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/http.errorResponse'
          description: Request Timeout
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/http.errorResponse'
          description: Conflict
        "413":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/http.errorResponse'
          description: Request Entity Too Large
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/http.RejectedContentResponse'
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/http.errorResponse'
          description: Internal Server Error
      summary: ReplaceContent handles PUT /internal/file/{id}/content (store-and-link)
      tags:
        - /internal
...
  /rest/storage/file/{id}:
    get:
      operationId: PublicHandler.ServeDocument
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: uuid
            type: string
      responses:
        "200":
          content:
            application/octet-stream:
              schema:
                format: binary
                type: string
          description: OK
        "304":
          description: Not Modified
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/http.errorResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/http.errorResponse'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/http.errorResponse'
          description: Internal Server Error
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/http.errorResponse'
          description: Service Unavailable
      summary: ServeDocument handles GET /rest/storage/file/{id} (and the /rest/storage/document/{id} back-compat alias)
      tags:
        - /rest/storage